pyproject.toml
[tool.poetry]
name = "chaebot"
version = "1.3.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
aiohttp = "^3.8.4"
"discord.py" = "^2.1.0"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
gunicorn = { extras = ["gevent"], version = "^20.1.0" }
python-twitter = "^3.5"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
tqdm = "^4.65.0"
whitenoise = { extras = ["brotli"], version = "^6.5.0" }
dj-database-url = "^2.0.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
watchdog = "^3.0.0"
click = "^8.1.3"
psycopg2-binary = "^2.9.6"
google-cloud-firestore = "^2.11.1"
firebase-admin = "^6.1.0"

[tool.poetry.group.dev.dependencies]
google-cloud-secret-manager = "^2.15.0"
ruff = "^0.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
ignore-init-module-imports = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E203", "E266", "E501", "E402"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["pydantic.validator"]
